+++ lib/src/math/algebra/interoperability/file_helper/file_io.dart
@@ -1,6 +1,6 @@
 export 'file_io_stub.dart'
     if (dart.library.html) 'file_io_web.dart'
-    if (dart.library.io) 'file_io_desktop.dart';
+    if (dart.library.io) 'file_io_other.dart';
 
 /// A class to provide file input/output operations.
 ///